Move navmesh reading
