+++ src/mesh_util.rs
@@ -1,0 +1,23 @@
+use bevy::prelude::*;
+use bevy::render::mesh::PrimitiveTopology;
+
+pub fn get_mesh<'a>(
+    children: &'a Children,
+    meshes: &'a Assets<Mesh>,
+    mesh_handles: &'a Query<&Handle<Mesh>>,
+) -> (Entity, &'a Mesh) {
+    let entity_handles: Vec<_> = children
+        .iter()
+        .filter_map(|entity| mesh_handles.get(*entity).ok().map(|mesh| (*entity, mesh)))
+        .collect();
+    assert_eq!(
+        entity_handles.len(),
+        1,
+        "Collider must contain exactly one mesh, but found {}",
+        entity_handles.len()
+    );
+    let (entity, mesh_handle) = entity_handles.first().unwrap();
+    let mesh = meshes.get(mesh_handle).unwrap();
+    assert_eq!(mesh.primitive_topology(), PrimitiveTopology::TriangleList);
+    (*entity, mesh)
+}
